Fix the hacked sleep delay.
This fancies up the tick return so that children can indicate they're waiting and be told when their other siblings are waiting. This should be pretty generalizable to other situations. I was trying to keep "Atlas" stuff out of the SensorModule related code.

Anyway, this hack was there because some modules seem to awaken when we're talking to others so this sleeps them all at once when we're done doing non-sleep things.