Major refactor to fix Custom queries.
These are handled by a separate task now and done a little more asynchronously
and so stuff had to move around. This also marks the beginning of the removal of
the unnecessary Sensor abstraction that just didn't turn into anything more
useful. I think for Atlas in particular we just focus on something concise.
There isn't a huge value in generalizing this stuff.